Lots of stuff here!
Alright, I'm going to start with the easy stuff. I noticed in my string that holds the API call, I had two '?'.  I got rid of one of them. I also saw that one of my parameters in the call was "api_key". That was also wrong, and I'm surprised it still worked.  I also created a struct for displaying a UIAlertController. 

The parse(:) function now returns an Optional instance of GIFModel. I wasn't able to specify the directory in path(:ofType:) because the function would only look in the top level directories. However it works just fine using the instance without it. I cleaned up the do/catch blocks in favor of using the nil coalescing operator. It looks cleaner and allows me to use all constants as opposed to mutable properties.

I added a parameter to fetchGIF(:completed:) that now takes a closure. The closure takes a Result type and I'm using it for error tracking. If I get an error, I can capture it and display an alert to the user. I'm also printing out the title of the GIF to the logs if the call is successful.

I think I caught everything, but I wouldn't be shocked if I missed something.